feat: Refactor the yahoo finance code into an adaptor class. Remove database file. Updated model folder to models
